customize

+++ database/migrations/2023_01_27_173531_create_orders_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->float("budget");
             $table->date("check_in");
             $table->date("check_out");
-            $table->integer("n_of_adults");
-            $table->integer("n_of_childeren")->default(0);
-            $table->integer("n_of_rooms");
+            $table->unsignedBigInteger("n_of_adults");
+            $table->unsignedBigInteger("n_of_childeren")->default(0);
+            $table->unsignedBigInteger("n_of_rooms");
         });
     }
